pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runnem"
version = "0.1.0"
description = "A service manager for managing multiple services in a project"
authors = [
    { name = "Daniel Murrell", email = "dsmurrell@gmail.com" }
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0.0",
    "requests>=2.31.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Repository = "https://github.com/dsmurrell/runnem.git"

[project.scripts]
runnem = "runnem.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "ruff>=0.4.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=runnem --cov-report=term-missing"

[tool.setuptools]
packages = ["runnem"]

[tool.ruff]
line-length = 300